feat(evals): replace in-page state machine with SvelteKit routes for eval config creation
Phase 1 of evals_v2_ui_fix: restructure the monolithic 748-line
create_eval_config page into proper SvelteKit routes, fixing broken
browser navigation (Back, refresh, deep-linking).

Architecture changes:
- Add +layout.svelte for shared data loading (eval, task, spec, models)
  with parallelized fetches via Promise.all
- Split picker (type selection cards) into index +page.svelte
- Add dynamic [eval_config_type] route for the builder form
- Extract EvalConfigBuilder container component (form, test pane, save
  logic, trust gate, save-without-testing dialog)
- Add shared context.ts (CreateEvalLayoutContext type + helper) and
  breadcrumbs.ts utility to eliminate duplication across routes
- Legacy ?config_type= query param redirect for backward compatibility
- Move eval_steps test to canonical $lib location, delete route-level
  duplicate
- Reorder eval types: LLM Judge first with "(recommended)" label

All 1177 tests pass, svelte-check 0 errors, lint/format/build clean.
